fix: escape chars processing, stripping
Remove unnecessary and incorrect whitespace stripping from the arguments
of some commands.
Fix escape character handling in arguments of some commands.
Previously `\x` was replaced with `x` for all values of x.
Now, `\n`, and `\\` are replaced with the newline character and `\`
character respectively, and rest appear as-is in the contents of those
arguments.